为QQ数据类型划分文件&为陌生人类添加toFriend方便函数
