+++ modules/QQDataTypes/HonorType.ts
@@ -1,0 +1,8 @@
+/**
+ * 荣誉类型
+ */
+export enum HonorType {
+    龙王 = "talkative",
+    群聊之火 = "performer",
+    快乐源泉 = "emotion"
+}
